Move WebAuth::Exception into a separate file, drop match method
Move WebAuth::Exception into its own *.pm file instead of including
it in WebAuth.pm.  Drop the match method, which was only used inside
the test suite.  Add some tests for the most important accessor
methods in the WebAuth::Exception class.

Add the missing WA_ERR_INVALID error code to the WebAuth Perl module
to allow for testing.

Move the Exporter and $VERSION information for WebAuth into a BEGIN
block so that it's safe for WebAuth and WebAuth::Exception to depend
on each other.

Change-Id: I2c28b81cfeb81df7accc7ea400696955c2ca8377
Reviewed-on: https://gerrit.stanford.edu/364
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>