Add WebLogin support for rejected authentication
Add a fourth parameter to WebKDCException holding additional data,
and fill that data in with the <userMessage> contents if present in
the WebKDC reply.  On an authentication rejected message, retrieve
that data and set it as a new err_html parameter to the error
template.  Update the default template to support err_html and use
that as the entire contents of the error box if set.

Change-Id: I55147b00812b040d3939a619289b6a2dbe85e668
Reviewed-on: https://gerrit.stanford.edu/423
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>