Set HttpOnly on WAS cookies by default
mod_webauth now sets the HttpOnly flag on all WebAuth session cookies
by default.  This can be turned off at the server or virtual host
level with the new WebAuthHttpOnly Apache directive.  (Although be
aware that the structure of the WebAuth cookies is an internal
implementation detail; if this directive is needed, the web site is
probably doing something unsupported.)

Change-Id: I1b9dcd54857b46362872b37c5d324881dbb04c4a
Reviewed-on: https://gerrit.stanford.edu/447
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>