Convert token decoding to use the new encoder
While we're at it, the reordering of operations means that we have
a full webauth_token struct to deal with and can therefore share
validation code with a minor tweak.  Eliminate all the now-redundant
validation code in the decoder and reuse the validation code in the
encoder, including moving over some helper routines to use with the
encoder as well.

Adjust the expected error messages in the test suite.

Also make the creation date in app tokens optional, since that's
what the specification says, although continue to generate that date
since that's been the behavior of recent releases.

Change-Id: I79ffa4c668ff4faf17048042763a937976f39f1d
Reviewed-on: https://gerrit.stanford.edu/501
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>