Add new WA_ERR_FILE_NOT_FOUND error
Add new WA_ERR_FILE_NOT_FOUND error, which replaces
WA_ERR_KEYRING_OPENREAD when the error is due to the file not existing.
Use it to distinguish between file not found errors and errors opening
an existing file in the keyring auto-update code and in the service
token cache code, where only file not found errors should be ignored.

Change-Id: Ice39c46d26548e09c672a5a12c36f65f1b7a5789
Reviewed-on: https://gerrit.stanford.edu/517
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>