WebLogin: set single sign-on cookies when displaying an error
In WebLogin, set single sign-on cookies if present even when
displaying an error.  This establishes single sign-on when errors are
returned after authentication, such as authentication rejected errors
from the user information service.  Without this behavior, if the
custom error sent the user to another page that also required
authentication, the user would have to log in again and may given up,
thinking that authentication was looping.

Change-Id: Iefda9771f3fd8340955ed7f5344c22c337acbfa6
Reviewed-on: https://gerrit.stanford.edu/519
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>