Stop using the legacy attribute code
Switch attribute encoding over to using a hand-rolled decoder that
builds a hash table of attribute names and values.  This removes the
last use in the tree of the legacy attribute code.

Remove the separate pool argument to attribute decoding.  This is
the first step in converting the Kerberos context stop using a
separate pool, which was a premature memory optimization.  Adjust
all callers.

Change-Id: Ib20fc15b15136dab59e7f9dfd49e0da1e6dc5948
Reviewed-on: https://gerrit.stanford.edu/577
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>