Check Perl scripts for strict, warnings, and syntax
Add a test script to check all Perl scripts in the distribution,
outside of the Perl directory, for use strict, use warnings, and
no syntax errors.  The Perl code in the perl directory will be
checked separately by a self-contained test suite.

Fix a few errors uncovered by this test.

Change-Id: Icd10f7a7794c9755fc00c4fa9a0069549a95ef50
Reviewed-on: https://gerrit.stanford.edu/581
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>