Make threshold for proxy token freshness configurable
Treat all proxy tokens as contributing to the session factors if
they are within a timeout specified as part of the WebKDC configuration,
and add a new Apache directive (WebKdcLoginTimeLimit) to control
that value, defaulting to the previous value of five minutes.

Update the force login code to honor that time limit rather than
requiring that a login token be provided.  This is in anticipation
of upcoming changes to support proxy authentication, where, for proxy
authentication to work with force login, it can't require a login
token be presented every time.

Change-Id: I4959ef9f72cca6136a3fc159a1ccb080d87a7c8a
Reviewed-on: https://gerrit.stanford.edu/612
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>