Support changing authz identity from WebLogin confirm page
Add a new runmode that redisplays the confirmation page after
changing the authz identity by resending the request and service
token with the new single sign-on cookie.  Include the RT and ST
in the parameters to the confirmation page.  Update the sample
confirmation page template to include a UI for changing the
authorization identity.

Change-Id: I88f61001f7763d7dceffac4b557441b72518b5eb
Reviewed-on: https://gerrit.stanford.edu/621
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>