Add WebLogin support for replay detection
When enabled, only one username and password authentication is
permitted with a given request token, and further authentications with
the same request token are rejected as replays.  This can protect
against an attacker using the back button in an abandoned browser
to replay the form submission on the WebLogin server.  This support
requires a memcached server be available for data storage and the Perl
modules Cache::Memcached and Digest::SHA.  The latter is available as
part of Perl since 5.9.3.

Change-Id: Ie2bf4e1d2c3f58ed111edd4125c2f79271ee9452
Reviewed-on: https://gerrit.stanford.edu/627
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>