Add authz logging recommendation to mod_webauth manual
When using authorization identities, Apache will log that identity
rather than the authentication identity.  It may be useful to also
log the authentication identity, and thankfully there's an unused
field for ident identity that can be reused.  Add a note to the
WebAuthTrustAuthzIdentity directive documentation suggesting a way
of modifying the Apache logging to do this.

Change-Id: Ibc1e6e85c3d8dd1f4e73c04be6fb54745300f708
Reviewed-on: https://gerrit.stanford.edu/630
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>