Fix encoding and decoding of ticket flags with Heimdal
Fix encoding of ticket flags with Heimdal Kerberos and tolerate the
old, incorrect encoding.  All previous versions of WebAuth, when built
with Heimdal, encoded the ticket flags on the wire with the flag bits
reversed (matching the in-memory Heimdal format).  Prior to this
version, flags would be lost when reading credentials encoded via MIT
Kerberos with Heimdal or vice versa.  As of this release, the portable
flag encoding used for ticket caches is used when writing credentials
with both MIT and Heimdal, and the flag order is detected when
decoding credentials and fixed if necessary.  If you use delegated
credentials and link with Heimdal Kerberos, upgrade mod_webauth prior
to upgrading the WebKDC to ensure the ticket flags are conveyed
correctly.

Change-Id: I058c573b210b0b1b26815e1c951ed222df7552a2
Reviewed-on: https://gerrit.stanford.edu/667
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>