Simplify Heimdal code for decoding flag bits
The logic that I originally copied from the Heimdal source was
excessively complex because it was manipulating a flag value that
was in network byte order rather than host byte order.  Our attribute
decoder guarantees that everything will be in host byte order before
we see it, so we don't have to manipulate the mask based on local
byte order.

Change-Id: I73a49c4e9ceecf88771925b9069b18e87d933e8c
Reviewed-on: https://gerrit.stanford.edu/669
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>