Add support for a WebLogin authenticate callback
Add a new authenticate callback to the WebLogin configuration.  If
this function is present in webkdc.conf, it will be called for every
user visit to WebLogin and may return the user's authentication
information or an empty list to defer to normal handling.  This can be
used to extract authentication information from the full WebLogin
environment; for example, it could map information about a successful
client-side certificate authentication to an authentication identity.

Change-Id: Ife843e431c021a4b7ea83b00090e8778830f4e23
Reviewed-on: https://gerrit.stanford.edu/716
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>