Add a new script to clean up generated module manuals
The XML format for documenting Apache modules, and the corresponding
build system to generate formatted HTML, is quite useful even outside
of Apache for documenting external modules.  However, the output of
the Apache documentation build system includes some Apache-specific
content, such as a comments section that only works on the Apache site
and a license (Apache 2.0) that may not be appropriate.

This script is used to post-process the formatted documentation for
the WebAuth Apache modules to remove those elements.  It should work
(but has not been tested with) Apache module documentation for other
third-party modules.

Change-Id: Id5df610e403b80c625965337742fd78006e2ace0
Reviewed-on: https://gerrit.stanford.edu/721
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>