Switch Perl build process to Module::Build
WebLogin and the WebAuth Perl bindings are now built with
Module::Build instead of ExtUtils::MakeMaker.  This should be
transparent to anyone not working with the source code, but
Module::Build and ExtUtils::CBuilder are now required to build the
WebLogin code.  They are included in Perl 5.10 or later and can be
installed separately for older versions of Perl.

Change-Id: I79f137bcc7e465e367c4c0fb423378fcd7ffc24a
Reviewed-on: https://gerrit.stanford.edu/726
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>