Ignore a subdirectory builddir in Perl tests
Give the Perl tests even more intelligence about directory manipulation
so that they can detect the case where the build directory is a
subdirectory of the source directory and where the path to the source
directory is given as .. instead of a full path.  In that case, don't
try to check syntax or coding style of files in the build directory.

Change-Id: I5055622cddb761efcc0ffa5f24e63ca8ae21c687
Reviewed-on: https://gerrit.stanford.edu/744
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>