Fix Perl builds when builddir != srcdir
Fix out-of-tree builds with --enable-webkdc.  Some of the logic to
ensure the Perl modules could build when the build directory was not
the source directory had not been tested for a while.

Not all the right directories were being created to copy over the
Perl source files, and the Perl build didn't have all the right -I
flags to pick up the portable/* header files.

Also simplify make distclean when builddir != srcdir to not bother
trying to remove the perl subdirectories.  make distclean in this
situation rarely tries hard to remove directories.

Change-Id: If8fd5d6d17fe51bf3e7c3e23bdca773bf5b75297
Reviewed-on: https://gerrit.stanford.edu/746
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>