Add protocol updates for persistent factors
Update the protocol specification to add factor tokens to the
<requestTokenRequest> and <requestTokenResponse> APIs and to define
a new webkdc-factor token type.  Add the cookie naming used to store
those tokens in the browser.  Update the user information service
API to pass already-established factors to the webkdc-userinfo
request and to allow the webkdc-validate response to set persistent
factors.

Change-Id: Iee0c776aa8774aa1df0499a77db88bb1729abad9
Reviewed-on: https://gerrit.stanford.edu/748
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>