Fix obscure bug in the Perl token-encode test
When checking raw encoding, we build the encoding we expect using
pack, but the old version of the test doesn't allow for the
possibility that the encoded time will happen to contain the byte
for an ASCII semicolon.  Add a function to wrap pack and double
the semicolon in that case.

(This actually randomly happened in one Debian package build.)

Change-Id: I9ad5cef3d3ecabe4e0a5ce3f19d6e48e17c2736c
Reviewed-on: https://gerrit.stanford.edu/751
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>