Remove limit on multivalued attributes in mod_webauthldap
Remove an arbitrary limit in mod_webauthldap on the number of values
from a multivalued LDAP attribute that are put in the environment.
Previous versions would only add the first 127 values, but there are
some cases where one may want to see more values than that.  This
opens the possibility of overflowing the allowed size of the
environment, but the maximum environment size is quite large on most
modern operating systems.

Change-Id: I2280929fe55d44f7c374c1362bcb2e91f37f8cf0
Reviewed-on: https://gerrit.stanford.edu/760
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>