Add webkdc-factor tokens
Add an implementation of webkdc-factor tokens, including the
corresponding Perl API and a test suite.  These tokens will hold
additional factors that aren't sufficient to authenticate the user
but that will be combined with the factors of an authentication to
set the factors of the resulting webkdc-proxy, id, and proxy tokens.

Change-Id: Ibbfc768a49bce954b8c8c5477db8e113752c337e
Reviewed-on: https://gerrit.stanford.edu/761
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>