Include webkdc-factor factors in a WebKDC authentication
Accept <factorToken> elements in <requestTokenRequest>.  Merge in
the factors they contribute provided that their expiration is less
than the expiration of the resulting webkdc-proxy token.  (This may
need to be reconsidered.)  Also take into account the persistent
factors returned by an OTP authentication and add them to the
webkdc-proxy token created by that authentication.

Fix some problems with promoting initial factors to session factors
when we only have a single webkdc-proxy token, and adjust the test
suite accordingly.

Change-Id: Icac42491a3279872fb920a2e7b3d9198358dffcb
Reviewed-on: https://gerrit.stanford.edu/787
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>