Get password expiration from WebKDC rather than remctl
The WebKDC passes back the password expiration time for an
authenticating user in the web response.  To avoid doubling up on the
remctl calls, we've pulled the remctl code that does the password
expiration lookups out of the WebLogin code and rely on the copy in the
web response.  Updated code to save it to the WebResponse object and to
clean tests and documentation for the new workings.

Change-Id: Ied39620a0b460d9cd4732021a20aa103fd951600
Reviewed-on: https://gerrit.stanford.edu/800
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>