Fix handling of equal authorization and authentication ids
If the user asserts an authorization identity equal to their
authentication identity, discard the authorization identity in the
WebKDC login process and continue as if they did not choose an
authorization identity.  This fixes a previously fatal error when the
user selects their default identity in WebLogin (if, for example, they
are trying to undo a previous choice of authorization identity).
Thanks to Benjamin Coddington for the report.

Change-Id: I1f2833b4e5b4de55e310a8312fbd7abe11e752a5
Reviewed-on: https://gerrit.stanford.edu/801
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>