Fix handling of equal authorization and authentication ids
If the user asserts an authorization identity equal to their
authentication identity, discard the authorization identity in the
WebKDC login process and continue as if they did not choose an
authorization identity.  This fixes a previously fatal error when the
user selects their default identity in WebLogin (if, for example, they
are trying to undo a previous choice of authorization identity).
Thanks to Benjamin Coddington for the report.

Reviewed-on: https://gerrit.stanford.edu/801
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>
(cherry picked from commit 3f08b0e50423a14585c87295b6171f4eab1b9abd)

Change-Id: I7d347da7ad51af60ee6a9e4d01da36a3019b477d
Reviewed-on: https://gerrit.stanford.edu/893
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>