Get device factor lifetime from the WebKDC
Rather than taking the device factor cookie lifetime from a config
setting (a temporary setup), it is now given from the WebKDC's response.
We create a new attribute on the WebResponse object to save the timeout
given by the WebKDC, then use that to set the cookie lifetime.  Also,
added a test set for the various device factor token pieces, both in
receiving a cookie to pass to the WebKDC, and receiving a setting from
the WebKDC to set a cookie.

Change-Id: Ia0dabfda05da541b8a66819a924d86ee0dcf03d1
Reviewed-on: https://gerrit.stanford.edu/803
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>