Updated the cookie pieces in WebResponse for more flexibility
The proxy_cookie and proxy_cookies functions in WebResponse have been
renamed cookie and cookies, to represent that they are now being used
for both proxy and device factor cookies.  They have also been modified
to take an expiration attribute (set to 0 if not passed), so that the
device factor expiration can be kept with the actual cookie data.  This
allows the possibility of more than one factor later on.

Change-Id: I654b60e3f6a904836806646cfda2bd1718675b33
Reviewed-on: https://gerrit.stanford.edu/805
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>