Fix obscure time bug in the Perl token-decode test
Similar to an earlier bug in token-encode, when checking raw encoding, we
build the encoding we expect using pack, but the old version of the test
doesn't allow for the possibility that the encoded time will happen to
contain the byte for an ASCII semicolon.  Add a function to wrap pack and
double the semicolon in that case.

(This actually randomly happened.)

Reviewed-on: https://gerrit.stanford.edu/821
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>
(cherry picked from commit 2e57152fb01fd3e484c3b25ff411d3dda0748e0c)

Change-Id: I6b3c993aaa1f8239d678b9edb1952ceb611e4061
Reviewed-on: https://gerrit.stanford.edu/896
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>