Fix obscure time bug in the Perl token-decode test
Similar to an earlier bug in token-encode, when checking raw encoding, we
build the encoding we expect using pack, but the old version of the test
doesn't allow for the possibility that the encoded time will happen to
contain the byte for an ASCII semicolon.  Add a function to wrap pack and
double the semicolon in that case.

(This actually randomly happened.)

Change-Id: Ie23b493204d99fd771c4555c4beac75e99068fa4
Reviewed-on: https://gerrit.stanford.edu/821
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>