Finish making the Perl module build self-contained
Remove the RRA::TAP::Automake library that's used to find files in
the main source tree during the test process.  Instead, copy all of
the files that the Perl tests need from the main tests into the
appropriate directories during the build and when setting up the
test suite.  Update the token-decode and token-encode tests and the
MANIFEST to expect those files to be present as part of the Perl
module.

Also fix a problem with removing the Perl *.o file between builds.

Change-Id: I1a3fb2e42320b96d9381e50240279dbfcdac6c9a
Reviewed-on: https://gerrit.stanford.edu/836
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>