Expire SSO cookies as well when the public computer flag is set
Expire the single sign-on cookies when the user has set that this is a
public computer.  As the cookies are needed during the login process, we
only want to expire at the end of the process.  If there is no confirm
page, we simply remove the cookies during the redirect to the WAS.  If
there is a confirm page, we clear the cookie while displaying the
confirm page.

Change-Id: I36c9a843f640ab54a3e746c668f641d4c9f65306
Reviewed-on: https://gerrit.stanford.edu/872
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>