Add an h factor
This indicates "human verification" and is used by a local support
organization to indicate off-line verification of identity so that
the user can get a multifactor factor.  Treat it as another class of
authentication for the purposes of synthesizing multifactor.

Add some protocol documentation for how implementations can choose
to synthesize multifactor factors.

Change-Id: I6e41fd6019928b2aa9adf1fb3a81a7e615e9c708
Reviewed-on: https://gerrit.stanford.edu/874
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>