Refactor WebLogin cookie handling to properly expire cookies
The cookie handling code in WebLogin was only iterating through
the cookies returned by the WebKDC and therefore only acting on
those, and ignoring any other cookies that may be set in the
browser.  This caused problems when clearing webauth_wpt and
webauth_wft cookies in circumstances where the WebKDC was not
setting new ones, such as after account lockout, because those
cookies weren't seen in the WebKDC response and therefore the code
to handle expiring them never triggered.

Build a combined list of cookies set by either the browser or the
WebKDC and iterate through all of them so that any cookie set by
the browser will be visible and can trigger expiration handling.

As a side effect, this will more correctly handle refreshing the
lifetime of the remote user cookie.

Change-Id: I2f54e4b90916b09171bc87b7b9184fea38a92fef
Reviewed-on: https://gerrit.stanford.edu/886
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>