Added warning for when long-lived multifactor about to expire
Added a new setting in WebKDC::Config to track a time period that we
want to issue warnings for, whenever a device factor is in that period.
Updated the default confirm template to print a warning in that case,
and WebLogin.pm to push out that data to the confirm template and
disable bypassing the confirm page, whenever we're in that period.

Change-Id: Ie73adefae1bc4748f344a8ea03db5d0ec1f6b525
Reviewed-on: https://gerrit.stanford.edu/939
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>