Continue supporting <multifactor-required> in user info
The new correct method is to send <required-factors>, but if that
element is not included and <multifactor-required /> is, add a
required factor of "m" to preserve the prior desired behavior, for
backward compatibility with existing implementations.

This also requires some test suite updates, which are coming in a
subsequent commit.

Change-Id: Icdae174aa0668379c87cd20034f8ec67df414fe6
Reviewed-on: https://gerrit.stanford.edu/920
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>