Keep WA_PEC_LOGIN_FAILED for all OTP failures
Don't switch to LOGIN_REJECTED when there is a user message.  We
don't know for sure that the condition is more of a rejected
condition instead of an incorrect code error.  Eventually, we need
a way for the user information service to tell us that, but we don't
have that yet.

Change-Id: I25c84a48a5897deb79f20ded9f9791ecca03e794
Reviewed-on: https://gerrit.stanford.edu/922
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>