Add APLOG_USE_MODULE calls to all module source
Apache 2.4 and later use this to initialize per-source-file data
that allows per-module logging levels to be implemented.  Use the
macro in all module source files and remove the previous extern
declarations of the module_data struct, which APLOG_USE_MODULE
handles declaring.  Define the macro on older versions of Apache
to only declare the module data in portable/apache.h.

Change-Id: Ic8f09abadb4b213f666935aceaab8285ea74fcdc
Reviewed-on: https://gerrit.stanford.edu/1004
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>