Move webkdc-proxy proxy subject checking into merge
When merging webkdc-proxy tokens, we need to check the subject
and proxy subject to ensure they match.  Move this code into the
webkdc-proxy merging function and return the new token error
code if they don't match.  Move the check that the merged
webkdc-proxy credential is a SSO credential out of this code and
into a separate check in the main login routine.

Change-Id: Ied62c19c05b9b56cef41d1d00e99d8f0aff1eb77
Reviewed-on: https://gerrit.stanford.edu/1012
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>