Move webkdc-factor token merging into a separate source file
Break this code out of webkdc-login.c and start a new source file
for the token merge functions.  Clean up the code somewhat and
constify it.

Add a new infrastructure for generic WebAuth test functions, the
first of which being a function to compare webkdc-factor tokens.
Use that to write a new data-driven test suite for webkdc-factor
merges.

Change-Id: I852c8c68beff288fc9ab198af648905cf93cfdc4
Reviewed-on: https://gerrit.stanford.edu/1015
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>