Start refactoring the WebKDC login test suite
Begin the long process of converting the test suite to be data-driven
instead of entirely in code.  Develop a new data structure to
represent the request and response for the WebKDC login process and
convert most of webkdc-login-t.c to use the new structure.

In the process, add new comparison functions for error, id, and proxy
tokens.

Change-Id: I3d800b43e5e8f9a617f44f8c60087def45db0795
Reviewed-on: https://gerrit.stanford.edu/1039
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>