Simplify the refactored WebKDC login test data
We can determine whether a login cancel token should be expected
by looking at the request, and it's always the same if present.
Move this out of the data.

The WebKDC always returns at most one webkdc-factor token, so stop
bothering with an array.  Instead, check that there is at most one
webkdc-factor token and check its contents directly.

Change-Id: I0baf2245c49d9e1b09f6169e2f5cb53e7e8e674d
Reviewed-on: https://gerrit.stanford.edu/1041
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>