Remove remaining derived data from WebKDC login tests
Clean up the rest of the WebKDC data-driven tests to remove all the
data that's derived from other data in the request or response.
Quite a bit of the response data is either derived from other
response data (duplicating encrypted data in the unencrypted section)
or is copied from the request (after decryption).

Change-Id: I302af3fbb39bd99ff1335c0683947fd39f8b811a
Reviewed-on: https://gerrit.stanford.edu/1042
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>