Add WebkDC login test infrastructure for Kerberos tests
Pass kerberos_config into all the test functions and support a simple
templating system that replaces various Kerberos-related placeholders
with information from the local Kerberos configuration.

Improve token expiration comparisons to allow the desired expiration
to be given as 0, which indicates that any expiration time in the future
is acceptable.

Fix webkdc-proxy token comparison when the proxy type is krb5.  In this
case, we don't want to literally compare the proxy data, just confirm
that it exists.  (Eventually, we may do something more complex, like
verifying that it is a Kerberos ticket cache.)

Add support for building templated login and proxy tokens.

Change-Id: I28236645ef30364730f48b4d79e6c2ead335cf24
Reviewed-on: https://gerrit.stanford.edu/1052
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>