Convert the rest of the webkdc-krb tests to the new structure
Generate a krb5.conf file for the webkdc-krb test so that we can
configure the local realm to be the configured test realm.  Without
this, authentication from unqualified principals doesn't work.

Update the generic WebAuth TAP add-on to correctly compare data
structures in id and proxy tokens when using krb5 authenticators,
where we can't provide test data in the wanted element.  Support
a really stupid templating mechanism for the login error message
for the message that includes the realm.  Insert webkdc-proxy tokens
into the array before login tokens so that login tokens override
webkdc-proxy tokens.  Increase the timestamp threshold to 10000 so
that an hour (60 * 60) isn't over the threshold.

Add a bunch of new tests for local realm handling, which wasn't
previously tested.

Change-Id: I8f3e637e23b17670071ed6c314dca218aedb33d7
Reviewed-on: https://gerrit.stanford.edu/1066
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>