Pass all tokens into webauth_webkdc_login encrypted
Change the API so that the webkdc-service and request tokens are
also passed in encrypted.  This means that all tokens are now passed
into the library encrypted, which will make the subsequent Perl
WebKDC implementation much simpler.

Currently, mod_webkdc still decrypts the webkdc-service and request
tokens so that it can do ACL checks, since the ACLs haven't been
moved into the library yet.

Change-Id: Ic7506af3e5bbb4143512d5072bb71a270e690788
Reviewed-on: https://gerrit.stanford.edu/1081
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>