Uniformly use s as the variable name for WebAuth status
We constantly use a variable to hold this all over the code, so may
as well use something short and consistent.  Change a few places
where we were using other terminlogy (err, code) to s, change all the
uses of status to s when they referred to WebAuth status, and change
status to code when it was from Kerberos or APR.  Also change the
internal webauth_context struct member to status from code for further
consistency.

Change-Id: Ic987ef1a7a605b08fbef9d2e0085e2829d78087a
Reviewed-on: https://gerrit.stanford.edu/1090
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>