Provide error context when encoding or decoding tokens
Use the new error context functionality to indicate the type of token
that is being encoded or decoded, or at least indicate that we're
decoding a token.  Simplify lots of other error handling now that we
don't have to try to get that information directly in the error message.
Remove some unnecessary verbiage from token crypto and attribute error
messages that is provided by context or by the default error message.

Change-Id: I45313f21a3de87a8e0d269ae507bd0fc0487e694
Reviewed-on: https://gerrit.stanford.edu/1096
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>