Use the return code for all WebKDC login processing
Previously, login processing distinguished between internal errors
(using the return status) and protocol errors (filling out the
login_error member of the response struct) and had to check both in
lots of places in the code.  Now that we've unified the code space,
just use the return status for everything.  We still fill out the
error code and message in the response in addition to returning a
correct status and setting the WebAuth context message.

As part of this work, fix a bunch of places where the new WebKDC
login code wasn't returning the correct rich protocol error codes,
and add more warning logging to ensure that we get rich error messages
in the WebKDC logs.

Change-Id: Ib8ae7b657c84814e54371c6726f9bef79c3740b7
Reviewed-on: https://gerrit.stanford.edu/1105
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>