Don't clear webkdc-factor tokens in WebLogin on initial login
We were always clearing the webkdc-factor token if the WebKDC didn't
send any cookies, but that's normal behavior on initial login or in
other circumstances where the WebKDC doesn't send a full response to
the login or doesn't have any webkdc-proxy tokens.  Only clear the
webkdc-factor cookie if we got at least one cookie back.

Change-Id: I18d4cbf7d8b726712a95eeafdce80abf4b50f03c
Reviewed-on: https://gerrit.stanford.edu/1110
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>