Revert forced authentication change, require a login token
Return to the behavior of requiring successful processing of a login
token to satisfy a forced authentication requirement, instead of
making it equivalent to a session factor requirement.  It seems more
useful to keep the concepts distinct, and the five minute window for
forced authentication never seemed quite right.

This means that forced authentication won't be usable with changing
the authorization subject from the WebLogin confirmation screen, since
the user would be forced to reauthenticate.

Change-Id: I300db6b77fd4b6814af9fd332a3c06721af4d885
Reviewed-on: https://gerrit.stanford.edu/1116
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>