Changed public_computer handling
We decided to change things to better represent how many places handle
this, with "Remember this computer".  To do so we renamed
public_computer remember_login and reversed the checks for how it works.
In order to give a proper fallback value since it's a checkbox, we
created $WebKDC::Config::REMEMBER_FALLBACK, so that you can give that
the opposite yes/no value for what the form checkbox is set to.

Change-Id: Icf113711cb76154a1480552bdc5716fa34857828
Reviewed-on: https://gerrit.stanford.edu/1157
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>