WebLogin: only expire factor tokens on confirm or redirect
The previous implementation would expire webkdc-factor tokens on
the first visit to the login page when displaying the login form,
since remember_login would not be set.  Only expire the cookies on
final redirect or display of the confirmation page.

Combine two related cases for webkdc-proxy tokens and update the
comment.

Change-Id: I5060fe28f5bd0513678ed31cef61053337b00b5c
Reviewed-on: https://gerrit.stanford.edu/1166
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>