WebLogin: Don't create undefined cookie objects
When WebLogin saw cookies that weren't webauth_* or internal
cookies, it would push an undefined value into the list of cookies.
This was ignored by the CGI module, but it's still not as clean as
desired.  Check whether we actually created a cookie object and
only add it to the list of cookies if we did.

Change-Id: Ife8bcc8a0a1a588076238681234ec46895687005
Reviewed-on: https://gerrit.stanford.edu/1167
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>