WebLogin: Always remember_login on single sign-on
If the request method to WebLogin is GET rather than POST, we're
in the initial page visit, possibly with single sign-on cookies.
In this case, we should always preserve any webkdc-proxy and
webkdc-factor cookies, even if displaying the confirmation page or
a final redirect, since we're doing single sign-on.  We only think
about deleting cookies if the method is POST, meaning the templates
have had a chance to set parameters like remember_login.

Change-Id: Ie1a2bd0a458f46c3914def0306e2209903bae1b1
Reviewed-on: https://gerrit.stanford.edu/1169
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>