WebLogin: Use full logic when setting remember_login
We can't just pass along whatever we saw as input, since we may
not have seen anything as input (if, for example, we bypassed the
initial login page because there was an existing proxy token and
went straight to the multifactor page).  Instead, for all the
pages that just reuse the setting as a hidden form element (the
multifactor and confirm pages), set it to whatever value we want
to use in the code, including the logic for special handling of
the defaults and the GET case.

Change-Id: I51cbd00032cb5b321fb41456090740a82e73ec04
Reviewed-on: https://gerrit.stanford.edu/1170
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>