WebLogin: Only destroy cookies on the confirmation page
If remember_login is set to no, WebLogin previously destroyed all
single sign-on and persistent factor cookies on either the
confirmation page or any redirect.  All redirects except one were
also confirmation pages, so this was mostly harmless.  But a redirect
is used to test cookie support if the test cookie is not set, so the
previous behavior would destroy device factor cookies if the test
cookie had to be set.

Only destroy single sign-on and factor cookies on the confirmation
page, not on other redirects.

Change-Id: I374772b550750f3b14e76429507617e13a0169e5
Reviewed-on: https://gerrit.stanford.edu/1181
Reviewed-by: Jon Robertson <jonrober@stanford.edu>
Tested-by: Russ Allbery <rra@stanford.edu>
Reviewed-by: Russ Allbery <rra@stanford.edu>